chore: mv sor_path to diff file
